+++ src/app/services/race.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { CacheService } from './cache.service';
 import { Race } from '../model/race';
-import * as firebase from 'firebase/app';
-import 'firebase/firestore';
+import firebase from 'firebase/compat/app';
+import 'firebase/compat/firestore';
 import { TimeService } from './time.service';
 
 @Injectable({